[sw/sca] Move libraries to sw/lib folder.
Moves all libraries to the sw/device/sca/lib folder and leaves the
SCA executables in sw/device/sca/.

Update util/licence-checker.hjson with the new path to the prng library.
