+++ sw/device/sca/lib/prng.c
