+++ sw/device/tests/sca/meson.build
@@ -5,7 +5,7 @@
 test('sca_prng_unittest', executable(
   'sca_prng_unittest',
   sources: [
-    meson.source_root() / 'sw/device/sca/prng.c',
+    meson.source_root() / 'sw/device/sca/lib/prng.c',
     'sca_prng_unittest.cc',
   ],
   dependencies: [
